% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcomp.R
\name{comp.test}
\alias{comp.test}
\title{Rejection probability test}
\usage{
comp.test (r, level, ng, ns, moys, varg, vars, method = "def.t")
}
\arguments{
\item{r}{the number of simulated data sets to use in finding each estimate.}

\item{level}{the significance level.}

\item{ng}{the number of groups in a data set.}

\item{ns}{the number of subjects per group.}

\item{moys}{the vector of means for the data sets.}

\item{varg}{the variance of group means.}

\item{vars}{the variance within a group.}

\item{method}{a character value specifying which t-test to apply. "def.t" (by default) indicates the standard t-test and "mod.t", the modified t-test.}
}
\value{
Returns a vector of estimated rejection probabilities for each mean.
}
\description{
estimates the rejection probability for a test at a given level.
}
\details{
The test is performed for various values of the overall mean, and the rejection probability estimated for each such value.
}
\author{
Sewanou Honfo \url{<honfosewanou@gmail.com>} and Romain Glèlè Kakaï \url{<glele.romain@gmail.com>}/ LABEF_07_2019
}
