% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcomp.R
\name{def.t}
\alias{def.t}
\title{Standard t test}
\usage{
def.t (X)
}
\arguments{
\item{X}{a \code{gen.data} object.}
}
\value{
Returns a list including:
\item{p.value}{the probability for the Standard two-sided t-test of the overall mean being zero.}
\item{moy.group}{the vector of means of values within each group.}
}
\description{
Performs the Standard t-test of mean = 0 for a one-way random effects model.
}
\author{
Sewanou Honfo \url{<honfosewanou@gmail.com>} and Romain Glèlè Kakaï \url{<glele.romain@gmail.com>}/ LABEF_07_2019
}
