% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcomp.R
\name{gen.data}
\alias{gen.data}
\title{Function for data generation}
\usage{
gen.data (ng, ns, moy, varg, vars)
}
\arguments{
\item{ng}{the number of groups.}

\item{ns}{the number of subject in each group. All groups have the same number of subjects.}

\item{moy}{the overall mean.}

\item{varg}{the variance of group means.}

\item{vars}{the variance of subjects within groups.
#'}
}
\value{
Returns a data frame with ng * ns rows, and two columns called "group" and "value", giving the group id for each subject (1 to ng), and the value for that subject.
}
\description{
generates data from a one-way random effects ANOVA model.
}
\author{
Sewanou Honfo \url{<honfosewanou@gmail.com>} and Romain Glèlè Kakaï \url{<glele.romain@gmail.com>} / LABEF_07_2019
}
