% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iwlsReg.R
\name{iwls.bnreg}
\alias{iwls.bnreg}
\title{Iterative Weighted Least Squares for logistic models}
\usage{
iwls.bnreg (y, X, n = NULL, intercept = FALSE, tol = 1e-07, max.iter = 1000)
}
\arguments{
\item{y}{response vector of observed proportions.}

\item{X}{model matrix of \code{p} covariates.}

\item{n}{the vector of binomial denominators. When NULL n is equal to a 1s vector.}

\item{intercept}{a logic value to indicate whether the intercept will be added directly by the function to the model matrix or not. The default value is FALSE i.e. add a vector of 1s to X.}

\item{tol}{the convergence tolerance criterion.}

\item{max.iter}{Maximum of iteration to be the limit if convergence is not attained. The default valueis 1000 iterations.}
}
\value{
Returns a list including the following:
\item{estimates}{the maximum likelihood estimates of the coefficients.}
\item{var}{the covariance matrix of coeeficients.}
\item{n.iter}{the number of iterations at convergence.}
}
\description{
Performs logisitc regressions for bionomial data.
}
\details{
The function performs the IWLS algorithm applied to binomial logistic regression (Fox, 2002)
}
\examples{

# Example on Mroz data from the package car
## Load packages and data
library(car)
data(Mroz)
str(Mroz)

## Convert non numeric variables to numeric
Mroz$lfp <- recode(Mroz$lfp, "'yes' = 1; 'no'= 0", as.factor = F)
Mroz$wc <- recode(Mroz$wc, "'yes' = 1; 'no'= 0", as.factor = F)
Mroz$hc <- recode(Mroz$hc, "'yes' = 1; 'no'= 0", as.factor = F)

## Run the binary logistic regression using the iterative weighted least squares estimation methods
attach(Mroz)
m.log <- iwls.bnreg(y = lfp, X = cbind(k5, k618, age, wc, hc, lwg, inc), n = NULL, intercept = FALSE, tol = 1e-07, max.iter = 1000)

## View the result
m.log$estimates
m.log$var
m.log$n.iter
}
\references{
Fox John (2002). An R and S-Plus companion to applied regression. Sage Publications.
}
\author{
Sewanou Honfo \url{<honfosewanou@gmail.com>} and Romain Glèlè Kakaï \url{<glele.romain@gmail.com>}/ LABEF_07_2019
}
