% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcomp.R
\name{mod.t}
\alias{mod.t}
\title{Modified t test}
\usage{
mod.t (X)
}
\arguments{
\item{X}{a \code{gen.data} object.}
}
\value{
Returns a list including:
\item{p.value}{the probability for the Modified two-sided t-test of the overall mean being zero.}
\item{moy.group}{the vector of means of values within each group.}
\item{vars.group}{the vector of variances of values within each group}
}
\description{
Performs the Modified t-test of mean = 0 for a one-way random effects model.
}
\author{
Sewanou Honfo \url{<honfosewanou@gmail.com>} and Romain Glèlè Kakaï \url{<glele.romain@gmail.com>}/ LABEF_07_2019
}
