% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcomp.R
\name{sim.test}
\alias{sim.test}
\title{A test on simulated data}
\usage{
sim.test (r, ng, ns, moy, varg, vars, method = "def.t")
}
\arguments{
\item{r}{the number of simulated data sets.}

\item{ng}{the number of groups in a data set.}

\item{ns}{the number of subjects per group.}

\item{moy}{the overall mean.}

\item{varg}{the variance of group means.}

\item{vars}{the variance within a group.}

\item{method}{a character value specifying which t-test to apply. "def.t"  (by default) indicates the standard t-test and "mod.t", the modified t-test.}
}
\value{
Returns a vector of p-values from the selected test.
}
\description{
applies a test to simulated one-way ANOVA data sets.
}
\author{
Sewanou Honfo \url{<honfosewanou@gmail.com>} and Romain Glèlè Kakaï \url{<glele.romain@gmail.com>}/ LABEF_07_2019
}
